fix: 로그아웃 버튼 수정
- 로그아웃 로고 삭제
- 토글 메뉴 아래에 로그아웃 버튼 배치
